requirements.txt
Django>=4.2,<4.3
djangorestframework
djangorestframework-simplejwt
